+++ packages/frontend/tests/performance/k6/helpers/getCookie.js
