+++ packages/frontend/tests/performance/k6/helpers/token.js
