+++ packages/frontend/tests/performance/k6/tests/browserFunctions.js
